save to database

+++ app/src/main/java/com/nick/geofence/Geofence.java
@@ -6,31 +6,31 @@ package com.nick.geofence;
 
 public class Geofence {
 
-    private long longitude;
-    private long latitude;
+    private double longitude;
+    private double latitude;
     private int radius; // meters
     private String wifiName;
 
-    public Geofence(long longt, long lat, int rad, String wifi) {
+    public Geofence(double longt, double lat, int rad, String wifi) {
         longitude = longt;
         latitude = lat;
         wifiName = wifi;
         radius = rad;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
